backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "python-dotenv>=1.0.1",
    "fastapi>=0.121.2",
    "pydantic>=2.12.4",
    "supabase>=2.24.0",
    "uvicorn[standard]>=0.38.0",
]
